break out solutions from muni
